+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Upgraded `load_modules`, `unload_modules` and `reload_modules` to the Client abstract interface.
 - `Component.make_loader` shorthand method for making a module loader and unloader for a component.
 - `tanjun.abc.ClientLoader` to make loaders more standard and easier to custom implement.
+- Command cooldowns.
 
 ### Changed
 - Renamed `Client.clear_commands` to `Client.clear_slash_commands`.